Fixed new warnings from merge

+++ src/restruct_poc/model/model.hpp
@@ -59,10 +59,10 @@ public:
     }
 
     /// Return a node by ID - const/read-only version
-    const Node& GetNode(int id) const { return *this->nodes_[id]; }
+    const Node& GetNode(size_t id) const { return *this->nodes_[id]; }
 
     /// Return a node by ID - non-const version
-    Node& GetNode(int id) { return *this->nodes_[id]; }
+    Node& GetNode(size_t id) { return *this->nodes_[id]; }
 
     /// Returns a reference to the nodes in the model (as vector of shared pointers)
     const std::vector<std::shared_ptr<Node>>& GetNodes() const { return this->nodes_; }
@@ -81,10 +81,10 @@ public:
     }
 
     /// Return a beam element by ID - const/read-only version
-    const BeamElement& GetBeamElement(int id) const { return *this->beam_elements_[id]; }
+    const BeamElement& GetBeamElement(size_t id) const { return *this->beam_elements_[id]; }
 
     /// Return a beam element by ID - non-const version
-    BeamElement& GetBeamElement(int id) { return *this->beam_elements_[id]; }
+    BeamElement& GetBeamElement(size_t id) { return *this->beam_elements_[id]; }
 
     /// Returns a reference to the beam elements in the model
     const std::vector<std::shared_ptr<BeamElement>>& GetBeamElements() const {
